Revert "Removed mail package"
